[BPF] Support wrapping BPF map structs into nested, single field structs
In Aya/Rust, BPF map definitions are nested in two wrapper types:

* Structs representing the given map type (`HashMap`, `RingBuf` etc.),
  which provide methods asocciated with that type. This way, we can
  define methods like `HashMap::get` or `RingBuf::reserve`, making the
  interation with maps more "rusty".
* `UnsafeCell`, which lets the Rust compiler know that such type is
  thread safe, can be defined as a static (global) variable and can be
  safely mutated. Linux kernel already guarantees the thread safety of
  map operations.

The type hierarchy of such nested map definitions usually looks like:

```
    pub struct HashMap<K, V, const M: usize, const F: usize = 0>(
        HashMapCell<K, V, M, F>
    );
    pub type HashMapCell<K, V, const M: usize, const F: usize = 0> =
        UnsafeCell<HashMapDef<K, V, M, F>>;
    pub struct HashMapDef<K, V, const M: usize, const F: usize = 0> {
        r#type: *const [i32; $t as usize],
        key: *const K,
        value: *const V,
        max_entries: *const [i32; M],
        map_flags: *const [i32; F],
    }
```

With the variable definition in the code of Aya users, like:

```
    #[link_section = ".maps"]
    static HASH_MAP: HashMap<u32, u32, 10> = HashMap::new();
```

being an equivalent of:

```
    struct {
        __uint(type, BPF_MAP_TYPE_HASH);
        __type(key, __u32);
        __type(value, __u32);
        __uint(max_entries, 10);
    } hash_map SEC(".maps");
```

Visiting such map requires visiting the fields:

```
  HASH_MAP -> __0 -> value
```

to get to the actual `HashMapDef` definition.

Before this change, the BPF backend was visiting only the fields of the
outer struct, without traversing them deeper. That caused problems with
maps using custom structs as keys and/or values, like:

```
    // Define custom structs for key and values.
    pub struct MyKey(u32);
    pub struct MyValue(u32);

    #[link_section = ".maps"]
    #[export_name = "HASH_MAP"]
    pub static HASH_MAP: HashMap<MyKey, MyValue, 10> = HashMap::new();
```

Because the `MyKey` and `MyValue` types didn't end up being actually
visited, and therefore ended up as a forward declaration in BTF:

```
    #30: <FWD> 'MyKey' kind:struct
    #31: <FWD> 'MyValue' kind:struct
```

Fix that by looking in the map definitions recursevely, if they turn out
to be single-field wrappers.

Fixes: #143361